Follow reindex_index API changes for PostgreSQL 9.5.
By commits [85b506] and [ecd222] of PostgreSQL,
the API reindex_index has gotten two more arguments.
The first is char persistence and the second is int
options. For pg_bulkload, the first should be checked
for each indexes, but it seems to be not necessary to
care about because this is related to REINDEX VERBOSE.